test(sec2): align 7 test schemas to the real BIGINT µ-unit money columns + guard
Phase-4-PRE. The Phase-3 finding "poolroyalty tests declare minted_amount BIGINT
while prod is DOUBLE" is INVERTED: migration 0082 (SEC-2) converts every family
mint table's minted_amount DOUBLE→BIGINT (0082:84-89) and traffic was born BIGINT
(0090), so prod is BIGINT. No conserved LENS/LXC column is DOUBLE in prod — no
SEC-2 regression; every prod DOUBLE money column is Tier-2/3 (USD/rate/score),
where float is deliberate.

The real divergences run the other way: because each integration test inlines
its own CREATE TABLE (no shared builder), stale pre-SEC-2 DOUBLE schemas survived
in isolated files that declare CONSERVED columns DOUBLE where prod is BIGINT —
passing only via pgx int64↔double coercion, so the suite never exercised the real
type. Align all seven to BIGINT µ-units:

- poolroyalty/held_clawback_realschema (minted_amount ×4 family) + a data_type
  assertion (was RED "double precision", now catches the divergence) + fixed the
  comments that mis-asserted prod is DOUBLE
- povi/node_harness, povi/slash_reputation, benchprobe/delivery_suppression
  (balances/ledger/povi_stakes/lxc)
- proxy/agent_allocator (lxc_*; + µLXC scale 50→50_000_000 to match prod default)
- mining/jsonb_simple_protocol (balances/ledger), earnverify/verify (lxc_purchases)

Authoritative guard: economy/sec2_realpg_integration_test asserts, against the
REAL migrated schema, BIGINT on all eight mint tables' minted_amount + the full
conserved set, AND double precision on the Tier-3 columns — both directions, so a
future conserved column shipping DOUBLE goes RED. Prod schema UNCHANGED (none
needed). The 0091 migration collision this branch originally also fixed was landed
independently by #302; rebased onto it.

+++ internal/mining/jsonb_simple_protocol_test.go
@@ -53,16 +53,16 @@ func TestLedgerMetadata_JSONBAcrossExecModes_Integration(t *testing.T) {
 				`DROP TABLE IF EXISTS lens_token_ledger`,
 				`DROP TABLE IF EXISTS lens_token_balances`,
 				`CREATE TABLE lens_token_balances (workspace_id TEXT PRIMARY KEY,
-					balance DOUBLE PRECISION NOT NULL DEFAULT 0,
-					lifetime_earned DOUBLE PRECISION NOT NULL DEFAULT 0,
-					lifetime_spent DOUBLE PRECISION NOT NULL DEFAULT 0,
-					updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW())`,
+					balance BIGINT NOT NULL DEFAULT 0,
+					lifetime_earned BIGINT NOT NULL DEFAULT 0,
+					lifetime_spent BIGINT NOT NULL DEFAULT 0,
+					updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW())`, // BIGINT µLENS (prod 0082)
 				// metadata NOT NULL DEFAULT '{}'::jsonb — the production shape
 				// (0019/0034). A NULL would violate it, so the fix must emit
 				// valid JSON text, never SQL NULL, for empty metadata.
 				`CREATE TABLE lens_token_ledger (id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-					workspace_id TEXT NOT NULL, amount DOUBLE PRECISION NOT NULL,
-					balance_after DOUBLE PRECISION NOT NULL, type TEXT NOT NULL, description TEXT,
+					workspace_id TEXT NOT NULL, amount BIGINT NOT NULL,
+					balance_after BIGINT NOT NULL, type TEXT NOT NULL, description TEXT,
 					metadata JSONB NOT NULL DEFAULT '{}'::jsonb, created_at TIMESTAMPTZ NOT NULL DEFAULT NOW())`,
 			} {
 				if _, err := pool.Exec(ctx, ddl); err != nil {
